feat: styling and configuration
