+++ pages/index.mdx
@@ -1,3 +1,5 @@
+import { Sponsors } from 'vocs/components'
+
 # Introduction 
 
 Starknet By Example is a collection of examples of how to use the Cairo programming language to create smart contracts on Starknet.
@@ -40,3 +42,5 @@ edition 2024_07
 sierra 1.6.0
 // [!include ~/.tool-versions]
 ```
+
+<Sponsors />
